3 items left to process fo the Countries database, the plan is to have this available always, but only show the "stored" countries when going offline.  There should eventually be a delta check to provide any updates to the countries data.  The installation should also include the existing snapshot of countries, and not have to process it the first time.
There is an efficiency in doing it this way, and the refresh or update to the country database can be detected in a micro service, or behave as the refresh of the web result set showing that the date it last changed.  Keeping our internal refresh of the last date that it was updated will allow us to simply check the data stamp and see if we need to refresh our CoreData.

I am going to continue to work on this for posterity, because I'd like to experiment with React Native is see if there are benefits and what are the drawbacks to doing it with React Native.